Refactoring

+++ get_or_update.sh
@@ -27,19 +27,20 @@ fi
 
 # check if config exists in /defaults, copy and configure if not
 if [ ! -e /defaults/nzbget.conf ]; then
-cp /nzbget/app/nzbget.conf /defaults/nzbget.conf
-sed -i \
-	-e "s#\(MainDir=\).*#\1/downloads#g" \
-	-e "s#\(ScriptDir=\).*#\1$\{MainDir\}/scripts#g" \
-	-e "s#\(WebDir=\).*#\1$\{AppDir\}/webui#g" \
-	-e "s#\(ConfigTemplate=\).*#\1$\{AppDir\}/webui/nzbget.conf.template#g" \
-/defaults/nzbget.conf
+	mkdir /defaults
+	cp /nzbget/app/nzbget.conf /defaults/nzbget.conf
+	sed -i \
+		-e "s#\(MainDir=\).*#\1/downloads#g" \
+		-e "s#\(ScriptDir=\).*#\1$\{MainDir\}/scripts#g" \
+		-e "s#\(WebDir=\).*#\1$\{AppDir\}/webui#g" \
+		-e "s#\(ConfigTemplate=\).*#\1$\{AppDir\}/webui/nzbget.conf.template#g" \
+	/defaults/nzbget.conf
 fi
 
 # check if config exists in /config, copy if not
-[[ ! -e /nzbget/config/nzbget.conf ]] && \
+if [[ ! -e /nzbget/config/nzbget.conf ]]; then
 	cp /defaults/nzbget.conf /nzbget/config/nzbget.conf
-
+fi
 
 
 # download the latest version of the nzbToMedia
